working triangle example

+++ examples/triangle/triangle.vert
@@ -8,10 +8,10 @@ vec2 positions[3] = vec2[](
     vec2(0.0, -0.5)
 );
 
-vec3 colors[3] = vec3[](
-    vec3(1.0, 0.0, 0.0),
-    vec3(0.0, 1.0, 0.0),
-    vec3(0.0, 0.0, 1.0)
+uvec3 colors[3] = uvec3[](
+    uvec3(1.0, 0.0, 0.0),
+    uvec3(0.0, 1.0, 0.0),
+    uvec3(0.0, 0.0, 1.0)
 );
 
 void main() {
